Use float for progress and hide when fully transparent
Use "float" for battery level everywhere (rather than an int * 1000 in some places). Don't show the animation when it would be invisible (i.e. when it would be fully transparent, alpha = 0)

+++ app/src/main/java/cn/vove7/energy_ring/monitor/DeviceMonitor.kt
@@ -16,5 +16,5 @@ interface DeviceMonitor {
 }
 
 interface MonitorListener {
-    fun onProgress(progress: Int)
+    fun onProgress(progress: Float)
 }
